requirements.txt
strands
strands-agents-tools
boto3
google-cloud-compute
google-cloud-iam
redis
